chore: big cleanup with improved shared code
