+++ internals/tanstack-query/src/index.ts
@@ -2,6 +2,7 @@ export { MutationKey, mutationKeyTransformer } from './components/MutationKey.ts
 export { QueryKey, queryKeyTransformer } from './components/QueryKey.tsx'
 export type { ParamsCasing, ParamsType, PathParamsType, Transformer } from './types.ts'
 export {
+  buildEnabledCheck,
   buildGroupParam,
   buildQueryKeyParams,
   resolveHeaderGroupType,
