+++ packages/plugin-vue-query/src/components/Mutation.tsx
@@ -4,7 +4,7 @@ import { functionPrinter } from '@kubb/plugin-ts'
 import { File, Function } from '@kubb/renderer-jsx'
 import type { KubbReactNode } from '@kubb/renderer-jsx/types'
 import type { PluginVueQuery } from '../types.ts'
-import { buildRequestConfigType, getComments, resolveErrorNames } from '../utils.ts'
+import { buildRequestConfigType, getComments, resolveErrorNames, wrapWithMaybeRefOrGetter } from '../utils.ts'
 
 type Props = {
   name: string
@@ -55,24 +55,7 @@ function buildMutationParamsNode(
 
   const mutationArgParamsNode = createMutationArgParams(node, { paramsCasing, resolver })
 
-  // Vue-query uses MutationObserverOptions instead of UseMutationOptions, and wraps params with MaybeRefOrGetter
-  const mutationArgWrapped = mutationArgParamsNode.params.map((param) => {
-    if (param.kind === 'ParameterGroup') {
-      return {
-        ...param,
-        properties: param.properties.map((property) => ({
-          ...property,
-          type: property.type ? ast.createParamsType({ variant: 'reference', name: `MaybeRefOrGetter<${printType(property.type)}>` }) : property.type,
-        })),
-      }
-    }
-
-    return {
-      ...param,
-      type: param.type ? ast.createParamsType({ variant: 'reference', name: `MaybeRefOrGetter<${printType(param.type)}>` }) : param.type,
-    }
-  })
-  const wrappedParamsNode = ast.createFunctionParameters({ params: mutationArgWrapped })
+  const wrappedParamsNode = wrapWithMaybeRefOrGetter(mutationArgParamsNode)
   const TRequestWrapped = wrappedParamsNode.params.length > 0 ? (declarationPrinter.print(wrappedParamsNode) ?? '') : ''
 
   return ast.createFunctionParameters({
@@ -90,21 +73,6 @@ function buildMutationParamsNode(
       }),
     ],
   })
-}
-
-function printType(typeNode: ast.ParamsTypeNode | undefined): string {
-  if (!typeNode) return 'unknown'
-  if (typeNode.variant === 'reference') return typeNode.name
-  if (typeNode.variant === 'member') return `${typeNode.base}['${typeNode.key}']`
-  if (typeNode.variant === 'struct') {
-    const parts = typeNode.properties.map((p) => {
-      const typeStr = printType(p.type)
-      const key = /^[a-zA-Z_$][a-zA-Z0-9_$]*$/.test(p.name) ? p.name : JSON.stringify(p.name)
-      return p.optional ? `${key}?: ${typeStr}` : `${key}: ${typeStr}`
-    })
-    return `{ ${parts.join('; ')} }`
-  }
-  return 'unknown'
 }
 
 export function Mutation({
