+++ packages/plugin-vue-query/src/generators/mutationGenerator.test.tsx
@@ -6,7 +6,7 @@ import { resolverTs } from '@kubb/plugin-ts'
 import { describe, test } from 'vitest'
 import { matchFiles } from '#mocks'
 import { mutationKeyTransformer } from '@internals/tanstack-query'
-import { queryKeyTransformer } from '../components/QueryKey.tsx'
+import { queryKeyTransformer } from '@internals/tanstack-query'
 import { resolverVueQuery } from '../resolvers/resolverVueQuery.ts'
 import type { PluginVueQuery } from '../types.ts'
 import { mutationGenerator } from './mutationGenerator.tsx'
